tests/factories/product_factory.py
import factory

from backend.apps.products.models import Product


class ProductFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Product

    name = factory.Sequence(lambda n: f"Product {n}")
    price = "99.99"
    stock = 10
